Refactor: style을 위한 util 함수 리팩토링
