+++ client/src/styles/theme.ts
@@ -1,5 +1,5 @@
 import { DefaultTheme } from 'styled-components';
-import { pixelToRem, createBreakPoint } from '~utils/style';
+import { pixelToRem, createBreakPoint } from '~utils/style-utils';
 
 export const colors = {
   primary: '#447A55',
